1558 lint fixes; minor template change and rename
